refactor all code modules/domain
refactor chuẩn code modules/domain
