+++ server/config.py
@@ -2,7 +2,7 @@ import os
 from pathlib import Path
 
 BASE_DIR = Path(__file__).resolve().parents[0]
-MODEL_DIR = Path(os.getenv("MODEL_DIR", BASE_DIR / "models_AI"))
+MODEL_DIR = Path(os.getenv("MODEL_DIR", BASE_DIR / "models"))
 DEFAULT_MODEL = os.getenv("DEFAULT_MODEL", "selfcontained_logreg.joblib")
 MODEL_PATH = MODEL_DIR / DEFAULT_MODEL
