+++ server/dependencies.py
@@ -1,0 +1,18 @@
+import os
+import jwt
+from fastapi import HTTPException, status, Request
+
+SUPABASE_JWT_SECRET = os.getenv("SUPABASE_JWT_KEY", "")
+AUDIENCE = "authenticated"
+
+def require_auth(request: Request) -> dict:
+    token = request.cookies.get("access_token")
+    if not token:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Missing token")
+    try:
+        payload = jwt.decode(token, SUPABASE_JWT_SECRET, algorithms=["HS256"], audience=AUDIENCE)
+        return payload
+    except jwt.ExpiredSignatureError:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token has expired")
+    except jwt.InvalidTokenError:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid token")
