+++ server/main.py
@@ -1,9 +1,14 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from server.core.db import lifespan
-from server.api.v1 import api_v1
+from server.database import lifespan
 import os
 from fastapi.staticfiles import StaticFiles
+from server.modules.docs.router import router as docs_router
+from server.modules.health.router import router as health_router
+from server.modules.users.router import router as users_router
+from server.modules.news.router import router as news_router
+from server.modules.ai.router import router as ai_router
+
 
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))  # ra ngoài app/
 STATIC_DIR = os.path.join(BASE_DIR, "server", "static")
@@ -25,4 +30,9 @@ app.add_middleware(
 )
 
 app.mount("/static", StaticFiles(directory=STATIC_DIR), name="static")
-app.include_router(api_v1, prefix="/api/v1")
+
+app.include_router(health_router, prefix="/api/v1")
+app.include_router(users_router, prefix="/api/v1")
+app.include_router(news_router, prefix="/api/v1")
+app.include_router(docs_router, prefix="/api/v1")
+app.include_router(ai_router, prefix="/api/v1")
