+++ server/modules/ai/router.py
@@ -1,7 +1,8 @@
 from fastapi import APIRouter, Depends, Header, HTTPException, Request
 from fastapi.responses import JSONResponse
-from server.schemas.ai_schema import MultipleNewsInputSchema, ClassificationResponseSchema, NewsAnalysisResponse, NewsAnalysisInput
-from server.services.ai_service import classify_news, analyze_news
+from server.modules.ai.schemas import MultipleNewsInputSchema, ClassificationResponseSchema, NewsAnalysisResponse, NewsAnalysisInput
+from server.modules.ai.service import classify_news, analyze_news
+from server.dependencies import require_auth
 from typing import List
 from dotenv import load_dotenv
 import requests
@@ -10,20 +11,20 @@ load_dotenv()
 N8N_WEBHOOK_URL = os.getenv("N8N_WEBHOOK_URL")
 
 
-router = APIRouter()
+router = APIRouter(prefix="/ai", tags=["AI"])
 
-@router.post("/classify_news", response_model=List[ClassificationResponseSchema])
-async def classify_news_route(news_data: MultipleNewsInputSchema, request: Request):
-    return classify_news(news_data.news, request)
+@router.post("/classify_news", response_model=List[ClassificationResponseSchema],dependencies=[Depends(require_auth)])
+async def classify_news_route(news_data: MultipleNewsInputSchema):
+    return classify_news(news_data.news)
 
-@router.post("/analyze-news", response_model=NewsAnalysisResponse)
-async def analyze_news_route(payload: NewsAnalysisInput, request: Request):
-    return analyze_news(payload, request)
+@router.post("/analyze-news", response_model=NewsAnalysisResponse, dependencies=[Depends(require_auth)])
+async def analyze_news_route(payload: NewsAnalysisInput):
+    return analyze_news(payload)
 
 @router.post("/chatbot")
 async def chatbot_route(
     text: str,
-    request: Request
+    access_token: str
 ):
     
     headers = {"Authorization": f"Bearer {access_token}"}
