+++ server/modules/ai/schemas.py
